requirements.txt
# Core dependencies
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
face-recognition==1.3.0
dlib==19.24.2

# Web framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Supabase
supabase==2.0.2
postgrest==0.13.0

# Database
sqlalchemy==2.0.21
psycopg2-binary==2.9.7

# Machine Learning
tensorflow==2.13.0
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pytest==7.4.2
pytest-cov==4.1.0

# GUI
pyqt5==5.15.9

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1
